refactor(web3signer): move the error, transport trait and decoding into their own modules
